+++ src/module-sanitychecker.cc
@@ -1,6 +1,6 @@
 /*
     Flexisip, a flexible SIP proxy server with media capabilities.
-    Copyright (C) 2010-2025 Belledonne Communications SARL, All rights reserved.
+    Copyright (C) 2010-2026 Belledonne Communications SARL, All rights reserved.
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU Affero General Public License as
@@ -16,66 +16,58 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include <flexisip/module.hh>
+#include "module-sanitychecker.hh"
 
-#include "agent.hh"
-#include "eventlogs/writers/event-log-writer.hh"
+#include "flexisip/flexisip-exception.hh"
+#include "flexisip/utils/sip-uri.hh"
 
 using namespace std;
 using namespace flexisip;
 
-class ModuleSanityChecker : public Module {
-	friend std::shared_ptr<Module> ModuleInfo<ModuleSanityChecker>::create(Agent*);
+unique_ptr<RequestSipEvent> ModuleSanityChecker::onRequest(unique_ptr<RequestSipEvent>&& ev) {
+	sip_t* sip = ev->getMsgSip()->getSip();
 
-public:
-	~ModuleSanityChecker() {
+	checkHeaders(sip);
+	if (sip->sip_request == nullptr || sip->sip_request->rq_url->url_host == nullptr) {
+		THROW_LINE(InvalidRequestError, "bad request URI");
 	}
+	// RFC 6665-3.1.2: https://datatracker.ietf.org/doc/html/rfc6665#section-3.1.2
+	if (sip->sip_request->rq_method == sip_method_subscribe && sip->sip_event == nullptr) {
+		THROW_LINE(InvalidRequestError, "bad request, no 'Event' header in SUBSCRIBE request");
+	}
+	return std::move(ev);
+}
 
-	unique_ptr<RequestSipEvent> onRequest(unique_ptr<RequestSipEvent>&& ev) override {
-		sip_t* sip = ev->getMsgSip()->getSip();
-
-		const char* error = checkHeaders(sip);
-		if (error) {
-			LOGD << "Rejecting request because of " << error;
-			ev->reply(400, error, SIPTAG_SERVER_STR(getAgent()->getServerString()), TAG_END());
-			return {};
+void ModuleSanityChecker::checkHeaders(sip_t* sip) {
+	if (sip->sip_via == nullptr) {
+		THROW_LINE(InvalidRequestError, "no via");
+	}
+	if (sip->sip_from == nullptr || sip->sip_from->a_url->url_host == nullptr || sip->sip_from->a_tag == nullptr) {
+		THROW_LINE(InvalidRequestError, "invalid from header");
+	}
+	if (sip->sip_to == nullptr || sip->sip_to->a_url->url_host == nullptr) {
+		THROW_LINE(InvalidRequestError, "invalid to header");
+	}
+	sofiasip::Url urlFrom{sip->sip_from->a_url};
+	auto fromParsingError = SipUri::hasParsingError(urlFrom);
+	if (fromParsingError) {
+		THROW_LINE(InvalidRequestError,
+		           "invalid URI in from header [" + urlFrom.str() + "]: " + fromParsingError.value());
+	}
+	sofiasip::Url urlTo{sip->sip_to->a_url};
+	auto toParsingError = SipUri::hasParsingError(urlTo);
+	if (toParsingError) {
+		THROW_LINE(InvalidRequestError, "invalid URI in to header [" + urlTo.str() + "]: " + toParsingError.value());
+	}
+	if (sip->sip_contact) {
+		if (sip->sip_contact->m_url->url_scheme == nullptr) {
+			THROW_LINE(InvalidRequestError, "invalid scheme in contact header");
 		}
-		if (sip->sip_request == nullptr || sip->sip_request->rq_url->url_host == nullptr) {
-			ev->reply(400, "Bad request URI", SIPTAG_SERVER_STR(getAgent()->getServerString()), TAG_END());
-			return {};
+		if (sip->sip_contact->m_url->url_scheme[0] != '*' && sip->sip_contact->m_url->url_host == nullptr) {
+			THROW_LINE(InvalidRequestError, "invalid contact header");
 		}
-		// RFC 6665-3.1.2: https://datatracker.ietf.org/doc/html/rfc6665#section-3.1.2
-		if (sip->sip_request->rq_method == sip_method_subscribe && sip->sip_event == nullptr) {
-			ev->reply(400, "Bad request, no 'Event' header in SUBSCRIBE request",
-			          SIPTAG_SERVER_STR(getAgent()->getServerString()), TAG_END());
-			return {};
-		}
-		return std::move(ev);
 	}
-
-	unique_ptr<ResponseSipEvent> onResponse(unique_ptr<ResponseSipEvent>&& ev) override {
-		// don't check our responses ;)
-		return std::move(ev);
-	}
-
-private:
-	ModuleSanityChecker(Agent* ag, const ModuleInfoBase* moduleInfo) : Module(ag, moduleInfo) {
-	}
-
-	const char* checkHeaders(sip_t* sip) {
-		if (sip->sip_via == nullptr) return "No via";
-		if (sip->sip_from == nullptr || sip->sip_from->a_url->url_host == nullptr || sip->sip_from->a_tag == nullptr)
-			return "Invalid from header";
-		if (sip->sip_to == nullptr || sip->sip_to->a_url->url_host == nullptr) return "Invalid to header";
-		if (sip->sip_contact) {
-			if (sip->sip_contact->m_url->url_scheme == nullptr) return "Invalid scheme in contact header";
-			if (sip->sip_contact->m_url->url_scheme[0] != '*' && sip->sip_contact->m_url->url_host == nullptr)
-				return "Invalid contact header";
-		}
-		return nullptr;
-	}
-	static ModuleInfo<ModuleSanityChecker> sInfo;
-};
+}
 
 ModuleInfo<ModuleSanityChecker> ModuleSanityChecker::sInfo(
     "SanityChecker",